refactor(Dropzone): create basic default story
